EY-2646: Treng snart grunnbeløp frå utanfor beregnings-modulen, så flyttar til model-lib
