+++ apps/etterlatte-beregning/src/main/kotlin/no/nav/etterlatte/grunnbeloep/GrunnbeloepService.kt
@@ -4,5 +4,5 @@ import no.nav.etterlatte.grunnbeloep.GrunnbeloepRepository
 import java.time.YearMonth
 
 class GrunnbeloepService(private val repository: GrunnbeloepRepository) {
-    fun hentGrunnbeloep(maaned: YearMonth) = repository.hentGjeldendeGrunnbeloep(maaned).grunnbeloep
+    fun hentGrunnbeloep(maaned: YearMonth) = repository.hentGjeldendeGrunnbeloep(maaned)
 }
